requirements.txt
streamlit
tensorflow
keras
numpy
pandas
scikit-learn
Pillow
